+++ csharp-netcore-modified/src/BungieNetApi/Model/DestinyComponentsProfilesDestinyProfileTransitoryTrackingEntry.cs
@@ -40,7 +40,7 @@ namespace BungieNetApi.Model
         /// <param name="activityHash">OPTIONAL - If this is tracking the status of a DestinyActivityDefinition, this is the identifier for that activity..</param>
         /// <param name="questlineItemHash">OPTIONAL - If this is tracking the status of a quest, this is the identifier for the DestinyInventoryItemDefinition that containst that questline data..</param>
         /// <param name="trackedDate">OPTIONAL - I&#39;ve got to level with you, I don&#39;t really know what this is. Is it when you started tracking it? Is it only populated for tracked items that have time limits?  I don&#39;t know, but we can get at it - when I get time to actually test what it is, I&#39;ll update this. In the meantime, bask in the mysterious data..</param>
-        public DestinyComponentsProfilesDestinyProfileTransitoryTrackingEntry(int locationHash = default(int), int itemHash = default(int), int objectiveHash = default(int), int activityHash = default(int), int questlineItemHash = default(int), DateTime trackedDate = default(DateTime))
+        public DestinyComponentsProfilesDestinyProfileTransitoryTrackingEntry(long locationHash = default(long), long itemHash = default(long), long objectiveHash = default(long), long activityHash = default(long), long questlineItemHash = default(long), DateTime trackedDate = default(DateTime))
         {
             this.LocationHash = locationHash;
             this.ItemHash = itemHash;
@@ -55,35 +55,35 @@ namespace BungieNetApi.Model
         /// </summary>
         /// <value>OPTIONAL - If this is tracking a DestinyLocationDefinition, this is the identifier for that location.</value>
         [DataMember(Name="locationHash", EmitDefaultValue=false)]
-        public int LocationHash { get; set; }
+        public long LocationHash { get; set; }
 
         /// <summary>
         /// OPTIONAL - If this is tracking the status of a DestinyInventoryItemDefinition, this is the identifier for that item.
         /// </summary>
         /// <value>OPTIONAL - If this is tracking the status of a DestinyInventoryItemDefinition, this is the identifier for that item.</value>
         [DataMember(Name="itemHash", EmitDefaultValue=false)]
-        public int ItemHash { get; set; }
+        public long ItemHash { get; set; }
 
         /// <summary>
         /// OPTIONAL - If this is tracking the status of a DestinyObjectiveDefinition, this is the identifier for that objective.
         /// </summary>
         /// <value>OPTIONAL - If this is tracking the status of a DestinyObjectiveDefinition, this is the identifier for that objective.</value>
         [DataMember(Name="objectiveHash", EmitDefaultValue=false)]
-        public int ObjectiveHash { get; set; }
+        public long ObjectiveHash { get; set; }
 
         /// <summary>
         /// OPTIONAL - If this is tracking the status of a DestinyActivityDefinition, this is the identifier for that activity.
         /// </summary>
         /// <value>OPTIONAL - If this is tracking the status of a DestinyActivityDefinition, this is the identifier for that activity.</value>
         [DataMember(Name="activityHash", EmitDefaultValue=false)]
-        public int ActivityHash { get; set; }
+        public long ActivityHash { get; set; }
 
         /// <summary>
         /// OPTIONAL - If this is tracking the status of a quest, this is the identifier for the DestinyInventoryItemDefinition that containst that questline data.
         /// </summary>
         /// <value>OPTIONAL - If this is tracking the status of a quest, this is the identifier for the DestinyInventoryItemDefinition that containst that questline data.</value>
         [DataMember(Name="questlineItemHash", EmitDefaultValue=false)]
-        public int QuestlineItemHash { get; set; }
+        public long QuestlineItemHash { get; set; }
 
         /// <summary>
         /// OPTIONAL - I&#39;ve got to level with you, I don&#39;t really know what this is. Is it when you started tracking it? Is it only populated for tracked items that have time limits?  I don&#39;t know, but we can get at it - when I get time to actually test what it is, I&#39;ll update this. In the meantime, bask in the mysterious data.
